requirements.txt
bidict==0.23.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
geventhttpclient==2.3.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.29.0
MarkupSafe==2.1.5
msgpack==1.0.8
psutil==6.0.0
python-engineio==4.9.1
python-socketio==5.11.2
pyzmq==26.0.3
redis==5.0.6
requests==2.32.3
simple-websocket==1.0.0
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
